requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-groq>=1.0.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
tavily-python>=0.3.0
streamlit>=1.50.0
